Fixed example.html to use current form field names & better error handling in getVisitData
I added more controls on when to update form fields with data. I also
added a fallback validation to rerun getVisitData 1 time after a
timeout if it can’t see any GA cookies in the client (__utmz, __utma,
etc…).

This should reduce the number of times we see WEB-unknown in form
submissions